Use RAS coords to guess hemisphere if voxel is 128
Previously, if the voxel coordinates were 128, the
program issued a warning that instructed the user to assign an
hemisphere manually for the corresponding electrode. With this change,
the program will look at the ras coordinates to try to determine the
hemisphere using the R coordinate (e.g. if the R coordinate is positive,
the hemisphere will be set to right, and vice-versa).