Allow hemisphere label 'ignore'
We allow this to notify the user that the program could not determine
the hemisphere without halting the program later. It warns the user to
use `set_hemisphere_manually` to specify the hemisphere themselves.

Parameters to the FreeSurferCoords class are introduced to control this
and the use of ras coordinates to help guess the hemisphere of the different
electrodes when the voxel method is inconclusive. Both default to false, and
the user should be aware that there is a possibility that electrodes close
to the midline could be misclassified.