Change behavior of `allow_ignore_hemi_label`
Will now control the creation of a `self.ignore` field within the coords
object. This field will contain a list of the channels whose hemisphere
could not be determined with the voxel. Note that this option labels the
electrode using the r coord to avoid any errors. Only the `self.ignore`
list contains the information about which electrodes could not be
determined using the voxel. Still defaults to false.